More permissive cse/constnat pooling aliasing condition

+++ torch/csrc/jit/passes/common_subexpression_elimination.cpp
@@ -25,13 +25,6 @@ void EliminateCommonSubexpression(
       continue;
     }
 
-    // since the graph outputs may be mutated after they are returned,
-    // don't introduce new aliasing
-    if (aliasDb.mayContainAlias(
-            node->outputs(), node->owningGraph()->outputs())) {
-      continue;
-    }
-
     if (!node->blocks().empty()) {
       // Traverse sub-blocks.
       for (auto block : node->blocks()) {
@@ -50,7 +43,15 @@ void EliminateCommonSubexpression(
 
     // Check for CSE opportunities in the parent block.
     auto parent_lookup = parent_lookup_fn(node);
+    auto g_out = node->owningGraph()->outputs();
     if (parent_lookup) {
+      // since the graph outputs may be mutated after they are returned,
+      // don't introduce new aliasing among graph outputs
+      if (aliasDb.mayContainAlias(node->outputs(), g_out) &&
+          aliasDb.mayContainAlias(parent_lookup->outputs(), g_out)) {
+        continue;
+      }
+
       node->replaceAllUsesWith(parent_lookup);
       it.destroyCurrent();
       continue;
@@ -61,6 +62,14 @@ void EliminateCommonSubexpression(
     if (!subit.second) {
       // Subexpression exists, replace the uses of node, and destroy it.
       auto existing = *subit.first;
+
+      // don't introduce new aliasing among graph outputs
+      if (aliasDb.mayContainAlias(
+              node->outputs(), node->owningGraph()->outputs()) &&
+          aliasDb.mayContainAlias(existing->outputs(), g_out)) {
+        continue;
+      }
+
       node->replaceAllUsesWith(existing);
       // Destroy the node.
       it.destroyCurrent();
